Allow customization of mapping between java and native types (wmeissner).  Type mapping is applied on function calls and structure read/write. Allow global specification of a library's native structure alignment. Allow more extensive customization of function name mapping. Provide default type/alignment/function mappings for w32 APIs (including stdcall decoration).  W32 examples have been updated to use the unicode version of functions by default (the w32.ascii system property may be set true to use ASCII versions instead). Support packed structures. Fix linux wide character support.
git-svn-id: https://svn.java.net/svn/jna~svn/trunk@128 2f8a963e-d2e4-e7d0-97bf-ccb7fcea9d80

+++ jnalib/src/com/sun/jna/examples/KeyboardUtils.java
@@ -157,7 +157,7 @@ public class KeyboardUtils {
                         int idx = code / 8;
                         int shift = code % 8;
                         if ((keys[idx] & (1 << shift)) != 0) {
-                            int sym = lib.XKeycodeToKeysym(dpy, code, 0);
+                            int sym = lib.XKeycodeToKeysym(dpy, (byte)code, 0);
                             if (sym == keysym)
                                 return true;
                         }
